GPU Acceleration & Multi-Threading & Path Fix
TensorFlow GPU Acceleration support for Windows. This is not enabled by default, but you can switch it on.

Dataset downloader and feature extraction will use Multi-Threading to speed up feature extraction on entire datasets, I/O was blocking on system saves and extraction, so by default it will use as many workers as your system has.