NPY Mapping fix
If you had more npy files than you did images (can happen with updating datasets) then the ordering of which images are displayed came out wrong. Because the ordering / id of the distance and the images was linked (Bad I know, but it was quick)

Now it will preserve the ordering by checking that both files exist in both folders (cache and dataset) to preserve the ordering.